Relocated `TargetActionEnvironmentMapper` to TuistGenerator
