+++ Sources/TuistCore/Graph/Mappers/TargetMapper.swift
@@ -3,22 +3,3 @@ import Foundation
 public protocol TargetMapping {
     func map(target: Target) throws -> (Target, [SideEffectDescriptor])
 }
-
-public class TargetActionEnvironmentMapper: TargetMapping {
-    let showEnvVarsInLog: Bool
-
-    public init(_ showEnvVarsInLog: Bool) {
-        self.showEnvVarsInLog = showEnvVarsInLog
-    }
-
-    public func map(target: Target) throws -> (Target, [SideEffectDescriptor]) {
-        var target = target
-        let actions: [TargetAction] = target.actions.map {
-            var action = $0
-            action.showEnvVarsInLog = showEnvVarsInLog
-            return action
-        }
-        target.actions = actions
-        return (target, [SideEffectDescriptor]())
-    }
-}
